Add functions from shapely to pygeos version
